Moves examples from js to jsx. Updates configs
