+++ examples/hackathon/postcss.config.cjs
@@ -1,7 +1,7 @@
 module.exports = {
   plugins: {
     tailwindcss: {
-      config: './tailwind.config.js',
+      config: './tailwind.config.cjs',
     },
     autoprefixer: {},
   },
