+++ examples/hackathon/tailwind.config.cjs
