+++ examples/realworld/src/client/Navbar.jsx
