+++ examples/realworld/src/client/WaspSourceHeader.jsx
