+++ examples/thoughts/src/client/MainPage.jsx
