+++ examples/thoughts/src/client/Tag.jsx
