+++ examples/thoughts/src/client/TagsSidebar.jsx
