+++ examples/thoughts/src/client/ThoughtsPage.jsx
