+++ examples/thoughts/src/client/TopNavbar.jsx
