+++ examples/thoughts/src/client/WaspSourceHeader.jsx
