+++ examples/thoughts/src/client/addWaspSourceHeader.jsx
@@ -1,12 +1,12 @@
 import React from 'react'
-import WaspSourceHeader from './WaspSourceHeader.js'
+import WaspSourceHeader from './WaspSourceHeader'
 
 const addWaspSourceHeader = (Component) => {
   return function AddHeader(props) {
     return (
       <>
-        <WaspSourceHeader name="Waspleau"/>
-        <Component { ...props } />
+        <WaspSourceHeader name="Thoughts" />
+        <Component { ...props }/>
       </>
     )
   }
