+++ examples/waspello/src/client/SignupPage.jsx
@@ -7,7 +7,7 @@ import login from '@wasp/auth/login.js'
 
 import EmailAndPassForm from './Auth/EmailAndPassForm'
 import GoogleAuthButton from './Auth/GoogleAuthButton'
-import addWaspSourceHeader from './addWaspSourceHeader.js'
+import addWaspSourceHeader from './addWaspSourceHeader'
 
 import mainLogo from './waspello-logo.svg'
 import './Signup.css'
