+++ examples/waspello/tailwind.config.cjs
