+++ examples/waspleau/src/client/addWaspSourceHeader.jsx
@@ -1,11 +1,11 @@
 import React from 'react'
-import WaspSourceHeader from './WaspSourceHeader.js'
+import WaspSourceHeader from './WaspSourceHeader'
 
 const addWaspSourceHeader = (Component) => {
   return function AddHeader(props) {
     return (
       <>
-        <WaspSourceHeader name="Waspello"/>
+        <WaspSourceHeader name="Waspleau"/>
         <Component { ...props } />
       </>
     )
